requirements.txt
cryptography==2.9.2
numpy==1.19.0
Pillow==7.2.0
tqdm==4.48.0